--- src/main/kotlin/nextstep/subway/path/domain/PathInterface.kt
@@ -1,5 +1,0 @@
-package nextstep.subway.path.domain
-
-interface PathInterface {
-    fun getShortestPath(paths: Paths): Map<String, List<String>>
-}
