--- src/main/kotlin/nextstep/subway/path/domain/Price.kt
@@ -1,8 +1,0 @@
-package nextstep.subway.path.domain
-
-enum class Price(val value: Int) {
-    BASIC(900),
-    ADD(100),
-    OVER_50KM_BASIC(800),
-    OUT(350)
-}
