--- src/main/kotlin/nextstep/subway/path/domain/TotalValue.kt
@@ -1,11 +1,0 @@
-package nextstep.subway.path.domain
-
-class TotalValue(private val values: List<Int>) {
-    fun get(type: String): List<Int> {
-        return if (type == "DISTANCE") {
-            values
-        } else {
-            values.asReversed()
-        }
-    }
-}
