+++ src/main/kotlin/nextstep/subway/path/shortest/domain/PathInterface.kt
@@ -1,0 +1,5 @@
+package nextstep.subway.path.shortest.domain
+
+interface PathInterface {
+    fun getShortestPath(paths: Paths): Map<String, List<String>>
+}
