+++ src/main/kotlin/nextstep/subway/path/shortest/domain/Price.kt
@@ -1,0 +1,8 @@
+package nextstep.subway.path.shortest.domain
+
+enum class Price(val value: Int) {
+    BASIC(900),
+    ADD(100),
+    OVER_50KM_BASIC(800),
+    OUT(350)
+}
