+++ src/main/kotlin/nextstep/subway/path/shortest/domain/TotalPrice.kt
@@ -1,0 +1,45 @@
+package nextstep.subway.path.shortest.domain
+
+import nextstep.subway.path.shortest.domain.Age.CHILD
+import nextstep.subway.path.shortest.domain.Age.TEENAGER
+import nextstep.subway.path.shortest.domain.Distance.OVER_10KM
+import nextstep.subway.path.shortest.domain.Distance.OVER_50KM
+
+object TotalPrice {
+    fun get(distance: Int, extraFare: Int = BASIC_EXTRA_FARE, age: Int = BASIC_AGE): Int {
+        val totalPrice = calculatePriceOfDistance(distance) + extraFare
+        return calculatePriceOfAge(totalPrice, age)
+    }
+
+    private fun calculatePriceOfAge(price: Int, age: Int): Int {
+        val totalPrice = when {
+            CHILD.itIn(age) -> {
+                CHILD.getPrice(price)
+            }
+            TEENAGER.itIn(age) -> {
+                TEENAGER.getPrice(price)
+            }
+            else -> {
+                price
+            }
+        }
+        return totalPrice + Price.OUT.value
+    }
+
+    private fun calculatePriceOfDistance(distance: Int): Int {
+        return when {
+            OVER_10KM.itIn(distance) -> {
+                OVER_10KM.getPrice(distance)
+            }
+            OVER_50KM.itIn(distance) -> {
+                OVER_50KM.getPrice(distance) + Price.OVER_50KM_BASIC.value
+            }
+            else -> {
+                Price.BASIC.value
+            }
+        }
+    }
+
+    private const val BASIC_AGE = 20
+    private const val BASIC_EXTRA_FARE = 0
+}
