+++ src/main/kotlin/nextstep/subway/path/shortest/dto/PathStationResponse.kt
@@ -1,0 +1,16 @@
+package nextstep.subway.path.shortest.dto
+
+import nextstep.subway.station.domain.Station
+import java.time.LocalDateTime
+
+class PathStationResponse(
+        val id: Long,
+        val name: String,
+        val createdAt: LocalDateTime?
+) {
+    companion object {
+        fun of(station: Station): PathStationResponse {
+            return PathStationResponse(station.id, station.name, station.createdDate)
+        }
+    }
+}
